requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
flake8==7.3.0
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
urllib3==2.5.0
